Solution: Foodlines complete and buggy
This version submits a close to working solution of the "foodlines"
problem: https://dmoj.ca/problem/lkp18c2p1

It needs a few runs through the debugger to smooth out the bugs.
It uses adapted implementations of `lowlevel/scanner/scanner.c` and
`lowlevel/scanner/twolines` in order to create a complete program.

The current bug is a segfault somewhere or other...

It first reads the two lines from stdin, using the stack to push / pop
values as they come in. The issue could be an off by one or something of
the like on the second line.

Signed-off-by: Adam McCartney <adam@mur.at>