Add two examples that use hashtables
This commit adds code for two solutions: "snowflakes" and
"compoundwords". Both are examples of programs that leverage the use of
hash tables to improve performance from n(n-1)/2 to O(n) and O(n^2) to
O(n) respectively. Basically we're targeting a linear time solution in
both cases. I think the trade off is that we are using slightly more
memory to initially setup the hash table, and so making a trade-off in
terms of using more space in memory in exchange for a notably quicker
execution time.

Signed-off-by: Adam McCartney <adam@mur.at>