Minimal binary tree lib implementation
Uses Knuth's algorithm for in order traversal.

_0_o_.

The library is essentially minimalist, provides only a single function
and assumes that all memory is handled by the caller.

This version also tries (and fails) to model the shape of the tree by
using a "level" variable to track the number of nodes visited.